aoj/ALDS1/ALDS1_13_C_test.py
from ALDS1_13_C import resolve
import sys
from io import StringIO
import unittest

class TestClass(unittest.TestCase):
    maxDiff = None
    def assertIO(self, input, output):
        stdout, stdin = sys.stdout, sys.stdin
        sys.stdout, sys.stdin = StringIO(), StringIO(input)
        resolve()
        sys.stdout.seek(0)
        out = sys.stdout.read()[:-1]
        sys.stdout, sys.stdin = stdout, stdin
        self.assertEqual(out, output)
    def test_入力例_1(self):
        input = """1 2 3 4
6 7 8 0
5 10 11 12
9 13 14 15"""
        output = """8"""
        self.assertIO(input, output)

if __name__ == "__main__":
    unittest.main()